dashboard/requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.20.0
google-cloud-bigquery==3.19.0
